Refactor metadata interfaces and remove unused files; add fetch polyfill for Node < 18
