+++ jest.config.js
@@ -1,7 +1,7 @@
 /** @type {import('jest').Config} */
 const config = {
     testEnvironment: 'node',
-    setupFiles: ['<rootDir>/test/setup.js'],
+    //setupFiles: ['<rootDir>/test/setup.js'],
 };
 
 module.exports = config;
