+++ src/interface/blockchainAction.ts
@@ -2,106 +2,81 @@ import { Abi, AbiStateMutability, AbiParameter } from './index';
 import { ContractFunctionName } from './index';
 import { ChainContext } from './chains';
 
-export interface UIParameter {
+// Input Types
+export type BaseInputType =
+    | 'text'
+    | 'number'
+    | 'boolean'
+    | 'email'
+    | 'url'
+    | 'datetime'
+    | 'textarea'
+    | 'address';
+// | 'bytes'    // Blockchain specific type
+// | 'hidden';  // Oculto en UI
+
+export type SelectionInputType = 'select' | 'radio';
+
+// Option for selects and radios
+export interface SelectOption {
     label: string;
+    value: string | number | boolean;
     description?: string;
-    minValue?: number;
-    maxValue?: number;
-    defaultValue?: any;
 }
 
-export interface BlockchainUIConfig {
-    title?: string;
-    description?: string;
-    parameters?: Record<string, UIParameter>;
+// Base parameter that all share
+export interface BaseParameter {
+    name: string; // Nombre del parámetro, debe coincidir con el ABI
+    label: string; // Etiqueta para mostrar en UI
+    description?: string; // Descripción - Ayuda
+    placeholder?: string; // Placeholder para inputs
+    required?: boolean; // Si es requerido
+    fixed?: boolean; // Si el valor es fijo, no editable
+    value?: any; // Valor por defecto
 }
 
-export interface BlockchainActionMetadata {
-    label: string;
-    address: `0x${string}`;
-    abi: Abi;
-    functionName: ContractFunctionName;
-    amount?: number; // Optional for DEVs to define the amount of the transaction - msg.value to be sent
-    paramsLabel?: string[]; // Optional for DEVs to define the label of the parameters
-    paramsValue?: (string | number | bigint | null | boolean)[]; // Optional for DEVs to define the value of the parameters
+// Parámetros standard (text, number, boolean, email, url, datetime, textarea)
+export interface StandardParameter extends BaseParameter {
+    type: BaseInputType;
+    minLength?: number; // Longitud mínima Texto
+    maxLength?: number; // Longitud máxima Texto
+    pattern?: string; // Para validaciones con regex
+    min?: number; // Para inputs numéricos y datetime
+    max?: number; // Para inputs numéricos y datetime
+}
+
+// Parámetro de selección (select, radio)
+export interface SelectParameter extends BaseParameter {
+    type: 'select';
+    options: SelectOption[];
+}
+
+// Parámetro de radio
+export interface RadioParameter extends BaseParameter {
+    type: 'radio';
+    options: SelectOption[];
+}
+
+// Tipo unión para todos los parámetros
+export type BlockchainParameter = StandardParameter | SelectParameter | RadioParameter;
+
+// Base Action Interface
+export interface BaseAction {
+    label: string; // Etiqueta para mostrar en UI
+    description: string; // Descripción - Ayuda
     chains: ChainContext; // [sourceChain, destinationChain | null]
-    ui?: BlockchainUIConfig;
 }
 
-// This interface is used internally to define the final blockchain action
+export interface BlockchainActionMetadata extends BaseAction {
+    label: string; // Etiqueta para mostrar en UI
+    address: `0x${string}`; // Dirección del contrato
+    abi: Abi; // ABI del contrato
+    functionName: ContractFunctionName; // Nombre de la función
+    amount?: number; // If function is payable, amount to send, set proper value as number - will be converted to WEI
+    params?: BlockchainParameter[]; // Array de parámetros configurados
+}
+
 export interface BlockchainAction extends BlockchainActionMetadata {
-    params: AbiParameter[];
-    blockchainActionType: AbiStateMutability;
+    abiParams: AbiParameter[]; // Processed ABI parameters
+    blockchainActionType: AbiStateMutability; // Function mutability
 }
-
-export interface TransferUIConfig {
-    title?: string;
-    description?: string;
-    successMessage?: string;
-    errorMessage?: string;
-    amountConfig?: {
-        label?: string;
-        description?: string;
-        placeholder?: string;
-        minValue?: number;
-        maxValue?: number;
-        step?: number;
-        errorMessage?: string;
-    };
-    addressConfig?: {
-        label?: string;
-        description?: string;
-        placeholder?: string;
-        errorMessage?: string;
-    };
-}
-
-export interface TransferAction {
-    label: string;
-    to?: `0x${string}`; // Recipient Address - IF NOT SENT Input will be shown
-    amount?: number; // Amount to transfer
-    chains: ChainContext; // [sourceChain, destinationChain | null]
-    // New optional UI configuration
-    ui?: TransferUIConfig;
-}
-
-// Plantillas predefinidas para configuraciones UI comunes
-export const UI_TEMPLATES = {
-    TOKEN_TRANSFER: {
-        title: 'Token Transfer',
-        description: 'Transfer tokens to another address',
-        successMessage: 'Transfer successful!',
-        errorMessage: 'Transfer failed. Please try again.',
-        parameters: {
-            to: {
-                label: 'Recipient Address',
-                description: "Enter the recipient's wallet address",
-                placeholder: '0x...',
-                pattern: '^0x[a-fA-F0-9]{40}$',
-                errorMessage: 'Invalid ethereum address',
-            },
-            amount: {
-                label: 'Amount',
-                description: 'Enter the amount to transfer',
-                placeholder: '0.0',
-                minValue: 0,
-                errorMessage: 'Amount must be greater than 0',
-            },
-        },
-    },
-    NFT_MINT: {
-        title: 'Mint NFT',
-        description: 'Mint a new NFT',
-        successMessage: 'NFT minted successfully!',
-        errorMessage: 'Minting failed. Please try again.',
-        parameters: {
-            tokenId: {
-                label: 'Token ID',
-                description: 'Enter the token ID to mint',
-                placeholder: '1',
-                minValue: 1,
-                errorMessage: 'Token ID must be greater than 0',
-            },
-        },
-    },
-} as const;
