+++ src/interface/transferAction.ts
@@ -1,0 +1,8 @@
+import { ChainContext } from "./chains";
+
+export interface TransferAction {
+    label: string;
+    to?: `0x${string}`;
+    amount?: number;
+    chains: ChainContext; // [sourceChain, destinationChain | null]
+}
